Compiler warning cleanup. In this chapter: silence narrowing conversion warnings and one shadowed argument.
* serf_private.h
  (SERF__CONV_assert): New debug-only assertion macro.
  (SERF__POSITIVE_TO_INT): Convert wider unsigned (or positive) values
   to int. Avoid typecasts that could yield a negative result.
  (SERF__SIGNED_TO_INT): Convert wider signed values to int.

* auth/auth.h
  (SERF_AUTH_assert): Remove, it's only use is gone.
* auth/auth.c
  (serf__encode_auth_header): Use SERF__POSITIVE_TO_INT.

* buckets/bwtp_buckets.c
  (parse_status_line):  Use SERF__POSITIVE_TO_INT.
* buckets/response_buckets.c
  (parse_status_line): And again.

* test/serf_bwtp.c (main): Just use casts for narrowing conversions,
   this is an example and must not use the private header.
* test/serf_spider.c (main): Likewise.
  (create_request): Rename the 'c' argument to avoid a shadowing warning.


git-svn-id: https://svn.apache.org/repos/asf/serf/trunk@1925997 13f79535-47bb-0310-9956-ffa450edef68

+++ test/serf_bwtp.c
@@ -517,7 +517,7 @@ int main(int argc, const char **argv)
 
         switch (opt_c) {
         case 'a':
-            srclen = strlen(opt_arg);
+            srclen = (int)strlen(opt_arg);
             enclen = apr_base64_encode_len(srclen);
             authn = apr_palloc(pool, enclen + 6);
             strcpy(authn, "Basic ");
@@ -538,7 +538,7 @@ int main(int argc, const char **argv)
             break;
         case 'n':
             errno = 0;
-            app_ctx.count = apr_strtoi64(opt_arg, NULL, 10);
+            app_ctx.count = (int)apr_strtoi64(opt_arg, NULL, 10);
             if (errno) {
                 printf("Problem converting number of times to fetch URL (%d)\n",
                        errno);
